+++ internal/app/app.go
@@ -41,8 +41,14 @@ type WAClient interface {
 
 	GetJoinedGroups(ctx context.Context) ([]*types.GroupInfo, error)
 	GetGroupInfo(ctx context.Context, jid types.JID) (*types.GroupInfo, error)
+	CreateGroup(ctx context.Context, req wa.CreateGroupRequest) (*types.GroupInfo, error)
 	SetGroupName(ctx context.Context, jid types.JID, name string) error
+	SetGroupTopic(ctx context.Context, jid types.JID, topic string) error
+	SetGroupAnnounce(ctx context.Context, jid types.JID, announce bool) error
+	SetGroupLocked(ctx context.Context, jid types.JID, locked bool) error
 	UpdateGroupParticipants(ctx context.Context, group types.JID, users []types.JID, action wa.GroupParticipantAction) ([]types.GroupParticipant, error)
+	GetGroupRequestParticipants(ctx context.Context, group types.JID) ([]types.GroupParticipantRequest, error)
+	UpdateGroupRequestParticipants(ctx context.Context, group types.JID, users []types.JID, action wa.GroupParticipantRequestAction) ([]types.GroupParticipant, error)
 	GetGroupInviteLink(ctx context.Context, group types.JID, reset bool) (string, error)
 	JoinGroupWithLink(ctx context.Context, code string) (types.JID, error)
 	LeaveGroup(ctx context.Context, group types.JID) error